Fixed few bugs.
Added a lot of checks for missing parameters, folders, files, etc.

Added preliminary help screen.

+++ trunk/cdroot/ultilex/bin/delete.bat
@@ -1,2 +1,16 @@
 @echo off
-rmdir /s /q ..\..\%1
+
+if (%1) == () goto missingParameter
+
+if exist ..\..\%1 rmdir /s /q ..\..\%1
+
+goto end
+
+:missingParameter
+echo.
+echo ERROR: Required parameter is missing. Use 'dist-get help' to get more information.
+echo.
+
+goto end
+
+:end
